{"schema":"decodex/commit/1","summary":"Split API docs L0 search route implementation into a child module after strict validation.","authority":"manual"}
